requirements.txt
python-telegram-bot==21.0
google-genai
supabase
python-dotenv